OPTIONAL/PARSING/lesson_5/requirements.txt
requests==2.27.1

beautifulsoup4==4.11.1
bs4==0.0.1

pymongo==4.1.1

python-dotenv==0.20.0

lxml==4.8.0

selenium==4.2.0